+++ consensus/module.go
@@ -1,16 +1,14 @@
 package consensus
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
+	"sync"
 
 	"github.com/pokt-network/pocket/consensus/leader_election"
 	typesCons "github.com/pokt-network/pocket/consensus/types"
+	"github.com/pokt-network/pocket/shared/codec"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
-	"github.com/pokt-network/pocket/shared/test_artifacts"
-	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/anypb"
 
 	consensusTelemetry "github.com/pokt-network/pocket/consensus/telemetry"
@@ -18,121 +16,133 @@ import (
 )
 
 const (
-	DefaultLogPrefix    = "NODE" // Just a default that'll be replaced during consensus operations.
+	DefaultLogPrefix    = "NODE" // TODO(#164): Make implicit when logging is standardized
 	ConsensusModuleName = "consensus"
 )
 
-var _ modules.ConsensusGenesisState = &typesCons.ConsensusGenesisState{}
-var _ modules.PacemakerConfig = &typesCons.PacemakerConfig{}
-var _ modules.ConsensusConfig = &typesCons.ConsensusConfig{}
-var _ modules.ConsensusModule = &ConsensusModule{}
+var (
+	_ modules.ConsensusModule       = &consensusModule{}
+	_ modules.ConsensusConfig       = &typesCons.ConsensusConfig{}
+	_ modules.ConsensusGenesisState = &typesCons.ConsensusGenesisState{}
+)
 
-// TODO(olshansky): Any reason to make all of these attributes local only (i.e. not exposed outside the struct)?
-// TODO(olshansky): Look for a way to not externalize the `ConsensusModule` struct
-type ConsensusModule struct {
+// TODO(#256): Do not export the `ConsensusModule` struct or the fields inside of it.
+type consensusModule struct {
 	bus        modules.Bus
 	privateKey cryptoPocket.Ed25519PrivateKey
-	consCfg    modules.ConsensusConfig
+
+	consCfg     modules.ConsensusConfig
+	consGenesis modules.ConsensusGenesisState
+
+	// m is a mutex used to control synchronization when multiple goroutines are accessing the struct and its fields / properties.
+	//
+	// The idea is that you want to acquire a Lock when you are writing values and a RLock when you want to make sure that no other goroutine is changing the values you are trying to read concurrently.
+	//
+	// Locking context should be the smallest possible but not smaller than a single "unit of work".
+	m sync.RWMutex
 
 	// Hotstuff
 	Height uint64
 	Round  uint64
 	Step   typesCons.HotstuffStep
-	Block  *typesCons.Block // The current block being voted on prior to committing to finality
+	Block  *typesCons.Block // The current block being proposed / voted on; it has not been committed to finality
 
-	HighPrepareQC *typesCons.QuorumCertificate // Highest QC for which replica voted PRECOMMIT
-	LockedQC      *typesCons.QuorumCertificate // Highest QC for which replica voted COMMIT
+	highPrepareQC *typesCons.QuorumCertificate // Highest QC for which replica voted PRECOMMIT
+	lockedQC      *typesCons.QuorumCertificate // Highest QC for which replica voted COMMIT
 
 	// Leader Election
 	LeaderId       *typesCons.NodeId
-	NodeId         typesCons.NodeId
-	ValAddrToIdMap typesCons.ValAddrToIdMap // TODO(design): This needs to be updated every time the ValMap is modified
-	IdToValAddrMap typesCons.IdToValAddrMap // TODO(design): This needs to be updated every time the ValMap is modified
+	nodeId         typesCons.NodeId
+	valAddrToIdMap typesCons.ValAddrToIdMap // TODO: This needs to be updated every time the ValMap is modified
+	idToValAddrMap typesCons.IdToValAddrMap // TODO: This needs to be updated every time the ValMap is modified
 
 	// Consensus State
-	appHash      string
+	lastAppHash  string // TODO: Always retrieve this variable from the persistence module and simplify this struct
 	validatorMap typesCons.ValidatorMap
 
 	// Module Dependencies
+	// TODO(#283): Improve how `utilityContext` is managed
 	utilityContext    modules.UtilityContext
 	paceMaker         Pacemaker
 	leaderElectionMod leader_election.LeaderElectionModule
 
-	logPrefix   string                                                  // TODO(design): Remove later when we build a shared/proper/injected logger
-	MessagePool map[typesCons.HotstuffStep][]*typesCons.HotstuffMessage // TODO(design): Move this over to the persistence module or elsewhere?
-	// TODO(andrew): Explain (or remove) why have an explicit `MaxBlockBytes` if we are already storing a reference to `consCfg` above?
-	// TODO(andrew): This needs to be updated every time the utility module changes this value. It can be accessed via the "application specific bus" (mimicking the intermodule interface in ABCI)
-	MaxBlockBytes uint64
+	// DEPRECATE: Remove later when we build a shared/proper/injected logger
+	logPrefix string
+
+	// TECHDEBT: Move this over to use the txIndexer
+	messagePool map[typesCons.HotstuffStep][]*typesCons.HotstuffMessage
 }
 
-func Create(configPath, genesisPath string, useRandomPK bool, configOptions ...func(*typesCons.ConsensusConfig)) (modules.ConsensusModule, error) {
-	cm := new(ConsensusModule)
-	c, err := cm.InitConfig(configPath)
-	if err != nil {
-		return nil, err
+func Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
+	return new(consensusModule).Create(runtimeMgr)
+}
+
+func (*consensusModule) Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
+	var m *consensusModule
+
+	cfg := runtimeMgr.GetConfig()
+	if err := m.ValidateConfig(cfg); err != nil {
+		return nil, fmt.Errorf("config validation failed: %w", err)
 	}
-	g, err := cm.InitGenesis(genesisPath)
-	if err != nil {
-		return nil, err
+	consensusCfg := cfg.GetConsensusConfig()
+
+	genesis := runtimeMgr.GetGenesis()
+	if err := m.ValidateGenesis(genesis); err != nil {
+		return nil, fmt.Errorf("genesis validation failed: %w", err)
 	}
-	cfg := c.(*typesCons.ConsensusConfig)
-	for _, o := range configOptions {
-		o(cfg)
-	}
-	genesis := g.(*typesCons.ConsensusGenesisState)
-	leaderElectionMod, err := leader_election.Create(cfg, genesis)
+	consensusGenesis := genesis.GetConsensusGenesisState()
+
+	leaderElectionMod, err := leader_election.Create(runtimeMgr)
 	if err != nil {
 		return nil, err
 	}
 
 	// TODO(olshansky): Can we make this a submodule?
-	paceMaker, err := CreatePacemaker(cfg)
+	paceMakerMod, err := CreatePacemaker(runtimeMgr)
 	if err != nil {
 		return nil, err
 	}
 
-	valMap := typesCons.ValidatorListToMap(genesis.Validators)
-	var privateKey cryptoPocket.PrivateKey
-	if useRandomPK {
-		privateKey, err = cryptoPocket.GeneratePrivateKey()
-	} else {
-		privateKey, err = cryptoPocket.NewPrivateKey(cfg.PrivateKey)
-	}
+	valMap := typesCons.ActorListToValidatorMap(consensusGenesis.GetVals())
+
+	privateKey, err := cryptoPocket.NewPrivateKey(consensusCfg.GetPrivateKey())
 	if err != nil {
 		return nil, err
 	}
 	address := privateKey.Address().String()
 	valIdMap, idValMap := typesCons.GetValAddrToIdMap(valMap)
 
-	m := &ConsensusModule{
+	paceMaker := paceMakerMod.(Pacemaker)
+
+	m = &consensusModule{
 		bus: nil,
 
-		privateKey: privateKey.(cryptoPocket.Ed25519PrivateKey),
-		consCfg:    cfg,
+		privateKey:  privateKey.(cryptoPocket.Ed25519PrivateKey),
+		consCfg:     cfg.GetConsensusConfig(),
+		consGenesis: genesis.GetConsensusGenesisState(),
 
 		Height: 0,
 		Round:  0,
 		Step:   NewRound,
 		Block:  nil,
 
-		HighPrepareQC: nil,
-		LockedQC:      nil,
+		highPrepareQC: nil,
+		lockedQC:      nil,
 
-		NodeId:         valIdMap[address],
+		nodeId:         valIdMap[address],
 		LeaderId:       nil,
-		ValAddrToIdMap: valIdMap,
-		IdToValAddrMap: idValMap,
+		valAddrToIdMap: valIdMap,
+		idToValAddrMap: idValMap,
 
-		appHash:      "",
+		lastAppHash:  "",
 		validatorMap: valMap,
 
 		utilityContext:    nil,
 		paceMaker:         paceMaker,
-		leaderElectionMod: leaderElectionMod,
+		leaderElectionMod: leaderElectionMod.(leader_election.LeaderElectionModule),
 
-		logPrefix:     DefaultLogPrefix,
-		MessagePool:   make(map[typesCons.HotstuffStep][]*typesCons.HotstuffMessage),
-		MaxBlockBytes: genesis.GetMaxBlockBytes(),
+		logPrefix:   DefaultLogPrefix,
+		messagePool: make(map[typesCons.HotstuffStep][]*typesCons.HotstuffMessage),
 	}
 
 	// TODO(olshansky): Look for a way to avoid doing this.
@@ -141,40 +151,7 @@ func Create(configPath, genesisPath string, useRandomPK bool, configOptions ...f
 	return m, nil
 }
 
-func (m *ConsensusModule) InitConfig(pathToConfigJSON string) (config modules.IConfig, err error) {
-	data, err := ioutil.ReadFile(pathToConfigJSON)
-	if err != nil {
-		return
-	}
-	// over arching configuration file
-	rawJSON := make(map[string]json.RawMessage)
-	if err = json.Unmarshal(data, &rawJSON); err != nil {
-		log.Fatalf("[ERROR] an error occurred unmarshalling the %s file: %v", pathToConfigJSON, err.Error())
-	}
-	// consensus specific configuration file
-	config = new(typesCons.ConsensusConfig)
-	err = json.Unmarshal(rawJSON[m.GetModuleName()], config)
-	return
-}
-
-func (m *ConsensusModule) InitGenesis(pathToGenesisJSON string) (genesis modules.IGenesis, err error) {
-	data, err := ioutil.ReadFile(pathToGenesisJSON)
-	if err != nil {
-		return
-	}
-	// over arching configuration file
-	rawJSON := make(map[string]json.RawMessage)
-	if err = json.Unmarshal(data, &rawJSON); err != nil {
-		log.Fatalf("[ERROR] an error occurred unmarshalling the %s file: %v", pathToGenesisJSON, err.Error())
-	}
-	// consensus specific configuration file
-	genesis = new(typesCons.ConsensusGenesisState)
-
-	err = json.Unmarshal(rawJSON[test_artifacts.GetGenesisFileName(m.GetModuleName())], genesis)
-	return
-}
-
-func (m *ConsensusModule) Start() error {
+func (m *consensusModule) Start() error {
 	m.GetBus().
 		GetTelemetryModule().
 		GetTimeSeriesAgent().
@@ -198,89 +175,55 @@ func (m *ConsensusModule) Start() error {
 	return nil
 }
 
-func (m *ConsensusModule) Stop() error {
+func (m *consensusModule) Stop() error {
 	return nil
 }
 
-func (m *ConsensusModule) GetModuleName() string {
+func (m *consensusModule) GetModuleName() string {
 	return ConsensusModuleName
 }
 
-func (m *ConsensusModule) GetBus() modules.Bus {
+func (m *consensusModule) GetBus() modules.Bus {
 	if m.bus == nil {
 		log.Fatalf("PocketBus is not initialized")
 	}
 	return m.bus
 }
 
-func (m *ConsensusModule) SetBus(pocketBus modules.Bus) {
+func (m *consensusModule) SetBus(pocketBus modules.Bus) {
 	m.bus = pocketBus
 	m.paceMaker.SetBus(pocketBus)
 	m.leaderElectionMod.SetBus(pocketBus)
 }
 
-func (m *ConsensusModule) loadPersistedState() error {
-	persistenceContext, err := m.GetBus().GetPersistenceModule().NewReadContext(-1) // Unknown height
-	if err != nil {
-		return nil
-	}
-	defer persistenceContext.Close()
-
-	latestHeight, err := persistenceContext.GetLatestBlockHeight()
-	if err != nil || latestHeight == 0 {
-		m.nodeLog("TODO: State sync not implement")
-		return nil
-	}
-
-	appHash, err := persistenceContext.GetBlockHash(int64(latestHeight))
-	if err != nil {
-		return fmt.Errorf("error getting block hash for height %d even though it's in the database: %s", latestHeight, err)
-	}
-
-	// TODO: Populate the rest of the state from the persistence module: validator set, quorum cert, last block hash, etc...
-	m.Height = uint64(latestHeight) + 1 // +1 because the height of the consensus module is where it is actively participating in consensus
-	m.appHash = string(appHash)
-
-	m.nodeLog(fmt.Sprintf("Starting node at height %d", latestHeight))
+func (*consensusModule) ValidateConfig(cfg modules.Config) error {
 	return nil
 }
 
-// TODO(discuss): Low priority design: think of a way to make `hotstuff_*` files be a sub-package under consensus.
-// This is currently not possible because functions tied to the `ConsensusModule`
-// struct (implementing the ConsensusModule module), which spans multiple files.
-/*
-TODO(discuss): The reason we do not assign both the leader and the replica handlers
-to the leader (which should also act as a replica when it is a leader) is because it
-can create a weird inconsistent state (e.g. both the replica and leader try to restart
-the Pacemaker timeout). This requires additional "replica-like" logic in the leader
-handler which has both pros and cons:
-	Pros:
-		* The leader can short-circuit and optimize replica related logic
-		* Avoids additional code flowing through the P2P pipeline
-		* Allows for micro-optimizations
-	Cons:
-		* The leader's "replica related logic" requires an additional code path
-		* Code is less "generalizable" and therefore potentially more error prone
-*/
-
-// TODO(olshansky): Should we just make these singletons or embed them directly in the ConsensusModule?
-type HotstuffMessageHandler interface {
-	HandleNewRoundMessage(*ConsensusModule, *typesCons.HotstuffMessage)
-	HandlePrepareMessage(*ConsensusModule, *typesCons.HotstuffMessage)
-	HandlePrecommitMessage(*ConsensusModule, *typesCons.HotstuffMessage)
-	HandleCommitMessage(*ConsensusModule, *typesCons.HotstuffMessage)
-	HandleDecideMessage(*ConsensusModule, *typesCons.HotstuffMessage)
+func (*consensusModule) ValidateGenesis(genesis modules.GenesisState) error {
+	return nil
 }
 
-func (m *ConsensusModule) HandleMessage(message *anypb.Any) error {
+func (m *consensusModule) GetPrivateKey() (cryptoPocket.PrivateKey, error) {
+	return cryptoPocket.NewPrivateKey(m.consCfg.GetPrivateKey())
+}
+
+func (m *consensusModule) HandleMessage(message *anypb.Any) error {
+	m.m.Lock()
+	defer m.m.Unlock()
 	switch message.MessageName() {
 	case HotstuffMessage:
-		var hotstuffMessage typesCons.HotstuffMessage
-		err := anypb.UnmarshalTo(message, &hotstuffMessage, proto.UnmarshalOptions{})
+		msg, err := codec.GetCodec().FromAny(message)
 		if err != nil {
 			return err
 		}
-		m.handleHotstuffMessage(&hotstuffMessage)
+		hotstuffMessage, ok := msg.(*typesCons.HotstuffMessage)
+		if !ok {
+			return fmt.Errorf("failed to cast message to HotstuffMessage")
+		}
+		if err := m.handleHotstuffMessage(hotstuffMessage); err != nil {
+			return err
+		}
 	case UtilityMessage:
 		panic("[WARN] UtilityMessage handling is not implemented by consensus yet...")
 	default:
@@ -290,42 +233,51 @@ func (m *ConsensusModule) HandleMessage(message *anypb.Any) error {
 	return nil
 }
 
-func (m *ConsensusModule) handleHotstuffMessage(msg *typesCons.HotstuffMessage) {
-	m.nodeLog(typesCons.DebugHandlingHotstuffMessage(msg))
-
-	// Liveness & safety checks
-	if err := m.paceMaker.ValidateMessage(msg); err != nil {
-		// If a replica is not a leader for this round, but has already determined a leader,
-		// and continues to receive NewRound messages, we avoid logging the "message discard"
-		// because it creates unnecessary spam.
-		if !(m.LeaderId != nil && !m.isLeader() && msg.Step == NewRound) {
-			m.nodeLog(typesCons.WarnDiscardHotstuffMessage(msg, err.Error()))
-		}
-		return
-	}
-
-	// Need to execute leader election if there is no leader and we are in a new round.
-	if m.Step == NewRound && m.LeaderId == nil {
-		m.electNextLeader(msg)
-	}
-
-	if m.isReplica() {
-		replicaHandlers[msg.Step](m, msg)
-		return
-	}
-
-	// Note that the leader also acts as a replica, but this logic is implemented in the underlying code.
-	leaderHandlers[msg.Step](m, msg)
+func (m *consensusModule) AppHash() string {
+	return m.lastAppHash
 }
 
-func (m *ConsensusModule) AppHash() string {
-	return m.appHash
-}
-
-func (m *ConsensusModule) CurrentHeight() uint64 {
+func (m *consensusModule) CurrentHeight() uint64 {
 	return m.Height
 }
 
-func (m *ConsensusModule) ValidatorMap() modules.ValidatorMap {
+func (m *consensusModule) ValidatorMap() modules.ValidatorMap { // TODO: This needs to be dynamically updated during various operations and network changes.
 	return typesCons.ValidatorMapToModulesValidatorMap(m.validatorMap)
+}
+
+// TODO(#256): Currently only used for testing purposes
+func (m *consensusModule) SetUtilityContext(utilityContext modules.UtilityContext) {
+	m.utilityContext = utilityContext
+}
+
+// TODO: Populate the entire state from the persistence module: validator set, quorum cert, last block hash, etc...
+func (m *consensusModule) loadPersistedState() error {
+	persistenceContext, err := m.GetBus().GetPersistenceModule().NewReadContext(-1) // Unknown height
+	if err != nil {
+		return nil
+	}
+	defer persistenceContext.Close()
+
+	latestHeight, err := persistenceContext.GetLatestBlockHeight()
+	if err != nil || latestHeight == 0 {
+		m.nodeLog("TODO: State sync not implemented yet")
+		return nil
+	}
+
+	appHash, err := persistenceContext.GetBlockHash(int64(latestHeight))
+	if err != nil {
+		return fmt.Errorf("error getting block hash for height %d even though it's in the database: %s", latestHeight, err)
+	}
+
+	m.Height = uint64(latestHeight) + 1 // +1 because the height of the consensus module is where it is actively participating in consensus
+	m.lastAppHash = string(appHash)
+
+	m.nodeLog(fmt.Sprintf("Starting node at height %d", latestHeight))
+	return nil
+}
+
+// HasPacemakerConfig is used to determine if a ConsensusConfig includes a PacemakerConfig without having to cast to the struct
+// (which would break mocks and/or pollute the codebase with mock types casts and checks)
+type HasPacemakerConfig interface {
+	GetPacemakerConfig() *typesCons.PacemakerConfig
 }