+++ p2p/raintree/peers_manager_test.go
@@ -6,9 +6,16 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/golang/mock/gomock"
 	"github.com/pokt-network/pocket/p2p/types"
+	"github.com/pokt-network/pocket/shared/crypto"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
+	modulesMock "github.com/pokt-network/pocket/shared/modules/mocks"
 	"github.com/stretchr/testify/require"
+)
+
+const (
+	serviceUrlFormat = "val_%d"
 )
 
 type ExpectedRainTreeNetworkConfig struct {
@@ -27,8 +34,6 @@ type ExpectedRainTreeMessageProp struct {
 	addrList string
 	targets  []ExpectedRainTreeMessageTarget
 }
-
-// IMPROVE(team): Looking into adding more tests and accounting for more edge cases.
 
 func TestRainTreeAddrBookUtilsHandleUpdate(t *testing.T) {
 	addr, err := cryptoPocket.GenerateAddress()
@@ -70,12 +75,11 @@ func TestRainTreeAddrBookUtilsHandleUpdate(t *testing.T) {
 			addrBook = append(addrBook, &types.NetworkPeer{Address: addr})
 			network := NewRainTreeNetwork(addr, addrBook).(*rainTreeNetwork)
 
-			err = network.processAddrBookUpdates()
-			require.NoError(t, err)
+			peersManagerStateView := network.peersManager.getNetworkView()
 
-			require.Equal(t, len(network.addrList), n)
-			require.Equal(t, len(network.addrBookMap), n)
-			require.Equal(t, int(network.maxNumLevels), testCase.numExpectedLevels)
+			require.Equal(t, len(peersManagerStateView.addrList), n)
+			require.Equal(t, len(peersManagerStateView.addrBookMap), n)
+			require.Equal(t, int(peersManagerStateView.maxNumLevels), testCase.numExpectedLevels)
 		})
 	}
 }
@@ -93,6 +97,11 @@ func BenchmarkAddrBookUpdates(b *testing.B) {
 		// {1000000000, 19},
 	}
 
+	// the test will add this arbitrary number of addresses after the initial initialization (done via NewRainTreeNetwork)
+	// this is to add extra subsequent work that -should- grow linearly and it's actually going to test AddressBook updates
+	// not simply initializations.
+	numAddressessToBeAdded := 1000
+
 	for _, testCase := range testCases {
 		n := testCase.numNodes
 		b.Run(fmt.Sprintf("n=%d", n), func(b *testing.B) {
@@ -100,12 +109,22 @@ func BenchmarkAddrBookUpdates(b *testing.B) {
 			addrBook = append(addrBook, &types.NetworkPeer{Address: addr})
 			network := NewRainTreeNetwork(addr, addrBook).(*rainTreeNetwork)
 
-			err = network.processAddrBookUpdates()
-			require.NoError(b, err)
+			peersManagerStateView := network.peersManager.getNetworkView()
 
-			require.Equal(b, len(network.addrList), n)
-			require.Equal(b, len(network.addrBookMap), n)
-			require.Equal(b, int(network.maxNumLevels), testCase.numExpectedLevels)
+			require.Equal(b, n, len(peersManagerStateView.addrList))
+			require.Equal(b, n, len(peersManagerStateView.addrBookMap))
+			require.Equal(b, testCase.numExpectedLevels, int(peersManagerStateView.maxNumLevels))
+
+			for i := 0; i < numAddressessToBeAdded; i++ {
+				newAddr, err := crypto.GenerateAddress()
+				require.NoError(b, err)
+				network.AddPeerToAddrBook(&types.NetworkPeer{Address: newAddr})
+			}
+
+			peersManagerStateView = network.peersManager.getNetworkView()
+
+			require.Equal(b, n+numAddressessToBeAdded, len(peersManagerStateView.addrList))
+			require.Equal(b, n+numAddressessToBeAdded, len(peersManagerStateView.addrBookMap))
 		})
 	}
 }
@@ -166,24 +185,32 @@ func TestRainTreeAddrBookTargetsTwentySevenNodes(t *testing.T) {
 }
 
 func testRainTreeMessageTargets(t *testing.T, expectedMsgProp *ExpectedRainTreeMessageProp) {
+	ctrl := gomock.NewController(t)
+	busMock := modulesMock.NewMockBus(ctrl)
+	consensusMock := modulesMock.NewMockConsensusModule(ctrl)
+	consensusMock.EXPECT().CurrentHeight().Return(uint64(1)).AnyTimes()
+	busMock.EXPECT().GetConsensusModule().Return(consensusMock).AnyTimes()
+
 	addrBook := getAlphabetAddrBook(expectedMsgProp.numNodes)
 	network := NewRainTreeNetwork([]byte{expectedMsgProp.orig}, addrBook).(*rainTreeNetwork)
-	network.processAddrBookUpdates()
+	network.SetBus(busMock)
 
-	require.Equal(t, strings.Join(network.addrList, ""), strToAddrList(expectedMsgProp.addrList))
+	peersManagerStateView := network.peersManager.getNetworkView()
 
-	i, found := network.getSelfIndexInAddrBook()
+	require.Equal(t, strings.Join(peersManagerStateView.addrList, ""), strToAddrList(expectedMsgProp.addrList))
+
+	i, found := network.peersManager.getSelfIndexInAddrBook()
 	require.True(t, found)
 	require.Equal(t, i, 0)
 
 	for _, target := range expectedMsgProp.targets {
-		addr, found := network.getFirstTargetAddr(uint32(target.level))
-		require.True(t, found)
-		require.Equal(t, addr, cryptoPocket.Address(target.left))
+		actualTargets := network.getTargetsAtLevel(uint32(target.level))
 
-		addr, found = network.getSecondTargetAddr(uint32(target.level))
-		require.True(t, found)
-		require.Equal(t, addr, cryptoPocket.Address(target.right))
+		require.True(t, shouldSendToTarget(actualTargets[0]))
+		require.Equal(t, actualTargets[0].address, cryptoPocket.Address(target.left))
+
+		require.True(t, shouldSendToTarget(actualTargets[1]))
+		require.Equal(t, actualTargets[1].address, cryptoPocket.Address(target.right))
 	}
 }
 
@@ -194,12 +221,14 @@ func getAlphabetAddrBook(n int) (addrBook types.AddrBook) {
 		if i >= n {
 			return
 		}
-		addrBook = append(addrBook, &types.NetworkPeer{Address: []byte{byte(ch)}})
+		addrBook = append(addrBook, &types.NetworkPeer{
+			ServiceUrl: fmt.Sprintf(serviceUrlFormat, i),
+			Address:    []byte{byte(ch)},
+		})
 	}
 	return
 }
 
 func strToAddrList(s string) string {
 	return hex.EncodeToString([]byte(s))
-
 }