+++ persistence/test/gov_test.go
@@ -2,10 +2,16 @@ package test
 
 import (
 	"encoding/hex"
-	"github.com/pokt-network/pocket/shared/modules"
 	"testing"
 
 	"github.com/stretchr/testify/require"
+)
+
+// TODO(#230): Remove these testing_artifacts when we nail down a design for parameter name sharing / owning
+const (
+	AppMaxChainsParamName            = "app_max_chains"
+	ServiceNodeMinimumStakeParamName = "service_node_minimum_stake"
+	ServiceNodeUnstakingBlocksOwner  = "service_node_unstaking_blocks_owner"
 )
 
 func TestInitParams(t *testing.T) {
@@ -22,13 +28,13 @@ func TestGetSetIntParam(t *testing.T) {
 
 	newMaxChains := 42
 
-	err = db.SetParam(modules.AppMaxChainsParamName, newMaxChains)
+	err = db.SetParam(AppMaxChainsParamName, newMaxChains)
 	require.NoError(t, err)
 
 	height, err := db.GetHeight()
 	require.NoError(t, err)
 
-	maxChains, err := db.GetIntParam(modules.AppMaxChainsParamName, height)
+	maxChains, err := db.GetIntParam(AppMaxChainsParamName, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newMaxChains, maxChains)
@@ -42,13 +48,13 @@ func TestGetSetStringParam(t *testing.T) {
 
 	newServiceNodeMinimumStake := "99999999"
 
-	err = db.SetParam(modules.ServiceNodeMinimumStakeParamName, newServiceNodeMinimumStake)
+	err = db.SetParam(ServiceNodeMinimumStakeParamName, newServiceNodeMinimumStake)
 	require.NoError(t, err)
 
 	height, err := db.GetHeight()
 	require.NoError(t, err)
 
-	serviceNodeMinimumStake, err := db.GetStringParam(modules.ServiceNodeMinimumStakeParamName, height)
+	serviceNodeMinimumStake, err := db.GetStringParam(ServiceNodeMinimumStakeParamName, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newServiceNodeMinimumStake, serviceNodeMinimumStake)
@@ -63,13 +69,13 @@ func TestGetSetByteArrayParam(t *testing.T) {
 	newOwner, err := hex.DecodeString("63585955783252764a6e576a5631647542486168426c63774e4655345a57617468545532637a6330516e4d5978575977674553537857644e4a6b4c7734575335416a65616c6d57494a47535364555933686d565a706e57564a6d6143526c54594248626864465a72646c624f646c59704a45536a6c6c52794d32527849545733566c6557464763745a465377466a57324a316157314562554a6c564b6c325470394753696c58544846474e786331567a70554d534a6c5335566d4c356f305157684663726c6b4e4a4e305931496c624a4e58537035554d4a7058564a705561506c3259484a47614b6c585a")
 	require.NoError(t, err)
 
-	err = db.SetParam(modules.ServiceNodeUnstakingBlocksOwner, newOwner)
+	err = db.SetParam(ServiceNodeUnstakingBlocksOwner, newOwner)
 	require.NoError(t, err)
 
 	height, err := db.GetHeight()
 	require.NoError(t, err)
 
-	owner, err := db.GetBytesParam(modules.ServiceNodeUnstakingBlocksOwner, height)
+	owner, err := db.GetBytesParam(ServiceNodeUnstakingBlocksOwner, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newOwner, owner)
@@ -84,13 +90,13 @@ func TestGetSetToggleIntFlag(t *testing.T) {
 	newMaxChains := 42
 
 	// insert with false
-	err = db.SetFlag(modules.AppMaxChainsParamName, newMaxChains, false)
+	err = db.SetFlag(AppMaxChainsParamName, newMaxChains, false)
 	require.NoError(t, err)
 
 	height, err := db.GetHeight()
 	require.NoError(t, err)
 
-	maxChains, enabled, err := db.GetIntFlag(modules.AppMaxChainsParamName, height)
+	maxChains, enabled, err := db.GetIntFlag(AppMaxChainsParamName, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newMaxChains, maxChains)
@@ -98,13 +104,13 @@ func TestGetSetToggleIntFlag(t *testing.T) {
 	require.Equal(t, false, enabled)
 
 	// toggle to true
-	err = db.SetFlag(modules.AppMaxChainsParamName, newMaxChains, true)
+	err = db.SetFlag(AppMaxChainsParamName, newMaxChains, true)
 	require.NoError(t, err)
 
 	height, err = db.GetHeight()
 	require.NoError(t, err)
 
-	maxChains, enabled, err = db.GetIntFlag(modules.AppMaxChainsParamName, height)
+	maxChains, enabled, err = db.GetIntFlag(AppMaxChainsParamName, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newMaxChains, maxChains)
@@ -121,26 +127,26 @@ func TestGetSetToggleStringFlag(t *testing.T) {
 	newServiceNodeMinimumStake := "99999999"
 
 	// insert with false
-	err = db.SetFlag(modules.ServiceNodeMinimumStakeParamName, newServiceNodeMinimumStake, false)
+	err = db.SetFlag(ServiceNodeMinimumStakeParamName, newServiceNodeMinimumStake, false)
 	require.NoError(t, err)
 
 	height, err := db.GetHeight()
 	require.NoError(t, err)
 
-	serviceNodeMinimumStake, enabled, err := db.GetStringFlag(modules.ServiceNodeMinimumStakeParamName, height)
+	serviceNodeMinimumStake, enabled, err := db.GetStringFlag(ServiceNodeMinimumStakeParamName, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newServiceNodeMinimumStake, serviceNodeMinimumStake)
 	require.Equal(t, false, enabled)
 
 	//toggle to true
-	err = db.SetFlag(modules.ServiceNodeMinimumStakeParamName, newServiceNodeMinimumStake, true)
+	err = db.SetFlag(ServiceNodeMinimumStakeParamName, newServiceNodeMinimumStake, true)
 	require.NoError(t, err)
 
 	height, err = db.GetHeight()
 	require.NoError(t, err)
 
-	serviceNodeMinimumStake, enabled, err = db.GetStringFlag(modules.ServiceNodeMinimumStakeParamName, height)
+	serviceNodeMinimumStake, enabled, err = db.GetStringFlag(ServiceNodeMinimumStakeParamName, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newServiceNodeMinimumStake, serviceNodeMinimumStake)
@@ -158,26 +164,26 @@ func TestGetSetToggleByteArrayFlag(t *testing.T) {
 	require.NoError(t, err)
 
 	// insert with false
-	err = db.SetFlag(modules.ServiceNodeUnstakingBlocksOwner, newOwner, false)
+	err = db.SetFlag(ServiceNodeUnstakingBlocksOwner, newOwner, false)
 	require.NoError(t, err)
 
 	height, err := db.GetHeight()
 	require.NoError(t, err)
 
-	owner, enabled, err := db.GetBytesFlag(modules.ServiceNodeUnstakingBlocksOwner, height)
+	owner, enabled, err := db.GetBytesFlag(ServiceNodeUnstakingBlocksOwner, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newOwner, owner)
 	require.Equal(t, false, enabled)
 
 	//toggle to true
-	err = db.SetFlag(modules.ServiceNodeUnstakingBlocksOwner, newOwner, true)
+	err = db.SetFlag(ServiceNodeUnstakingBlocksOwner, newOwner, true)
 	require.NoError(t, err)
 
 	height, err = db.GetHeight()
 	require.NoError(t, err)
 
-	owner, enabled, err = db.GetBytesFlag(modules.ServiceNodeUnstakingBlocksOwner, height)
+	owner, enabled, err = db.GetBytesFlag(ServiceNodeUnstakingBlocksOwner, height)
 	require.NoError(t, err)
 
 	require.Equal(t, newOwner, owner)