+++ shared/crypto/keys.go
@@ -7,7 +7,7 @@ import (
 	"log"
 )
 
-// TODO(discuss): Consider create a type for signature and having constraints for each type as well.
+// DISCUSS Consider create a type for signature and having constraints for each type as well.
 type Address []byte
 
 type PublicKey interface {