+++ utility/block.go
@@ -1,13 +1,9 @@
 package utility
 
 import (
-	"math/big"
-
-	typesCons "github.com/pokt-network/pocket/consensus/types" // TODO (andrew) importing consensus and persistence in this file?
-	typesGenesis "github.com/pokt-network/pocket/persistence/types"
 	"github.com/pokt-network/pocket/shared/modules"
-
 	typesUtil "github.com/pokt-network/pocket/utility/types"
+	"math/big"
 )
 
 /*
@@ -21,16 +17,6 @@ import (
 	rolls it back (not yet implemented) if fails. Like BeginBlock, EndBlock is an autonomous state
 	operation that executes at the end of every block.
 */
-
-// TODO(andrew): consolidate with `utility/types/actor.go`
-var (
-	UtilActorTypes = []typesUtil.UtilActorType{
-		typesUtil.UtilActorType_App,
-		typesUtil.UtilActorType_Node,
-		typesUtil.UtilActorType_Fish,
-		typesUtil.UtilActorType_Val,
-	}
-)
 
 func (u *UtilityContext) ApplyBlock(latestHeight int64, proposerAddress []byte, transactions [][]byte, lastBlockByzantineValidators [][]byte) ([]byte, error) {
 	u.LatestHeight = latestHeight
@@ -103,7 +89,7 @@ func (u *UtilityContext) GetAppHash() ([]byte, typesUtil.Error) {
 
 // HandleByzantineValidators handles the validators who either didn't sign at all or disagreed with the 2/3+ majority
 func (u *UtilityContext) HandleByzantineValidators(lastBlockByzantineValidators [][]byte) typesUtil.Error {
-	latestBlockHeight, err := u.GetLatestHeight()
+	latestBlockHeight, err := u.GetLatestBlockHeight()
 	if err != nil {
 		return err
 	}
@@ -129,7 +115,7 @@ func (u *UtilityContext) HandleByzantineValidators(lastBlockByzantineValidators 
 			if err != nil {
 				return err
 			}
-			if err = u.BurnActor(typesUtil.UtilActorType_Val, burnPercentage, address); err != nil {
+			if err = u.BurnActor(typesUtil.ActorType_Validator, burnPercentage, address); err != nil {
 				return err
 			}
 		} else if err := u.SetValidatorMissedBlocks(address, numberOfMissedBlocks); err != nil {
@@ -142,23 +128,29 @@ func (u *UtilityContext) HandleByzantineValidators(lastBlockByzantineValidators 
 func (u *UtilityContext) UnstakeActorsThatAreReady() (err typesUtil.Error) {
 	var er error
 	store := u.Store()
-	latestHeight, err := u.GetLatestHeight()
+	latestHeight, err := u.GetLatestBlockHeight()
 	if err != nil {
 		return err
 	}
-	for _, utilActorType := range typesUtil.ActorTypes {
+	for _, actorTypeInt32 := range typesUtil.ActorType_value {
 		var readyToUnstake []modules.IUnstakingActor
-		poolName := utilActorType.GetActorPoolName()
-		switch utilActorType {
-		case typesUtil.UtilActorType_App:
-			readyToUnstake, er = store.GetAppsReadyToUnstake(latestHeight, typesUtil.UnstakingStatus)
-		case typesUtil.UtilActorType_Fish:
-			readyToUnstake, er = store.GetFishermenReadyToUnstake(latestHeight, typesUtil.UnstakingStatus)
-		case typesUtil.UtilActorType_Node:
-			readyToUnstake, er = store.GetServiceNodesReadyToUnstake(latestHeight, typesUtil.UnstakingStatus)
-		case typesUtil.UtilActorType_Val:
-			readyToUnstake, er = store.GetValidatorsReadyToUnstake(latestHeight, typesUtil.UnstakingStatus)
-
+		actorType := typesUtil.ActorType(actorTypeInt32)
+		var poolName string
+		switch actorType {
+		case typesUtil.ActorType_App:
+			readyToUnstake, er = store.GetAppsReadyToUnstake(latestHeight, int32(typesUtil.StakeStatus_Unstaking))
+			poolName = typesUtil.PoolNames_AppStakePool.String()
+		case typesUtil.ActorType_Fisherman:
+			readyToUnstake, er = store.GetFishermenReadyToUnstake(latestHeight, int32(typesUtil.StakeStatus_Unstaking))
+			poolName = typesUtil.PoolNames_FishermanStakePool.String()
+		case typesUtil.ActorType_ServiceNode:
+			readyToUnstake, er = store.GetServiceNodesReadyToUnstake(latestHeight, int32(typesUtil.StakeStatus_Unstaking))
+			poolName = typesUtil.PoolNames_ServiceNodeStakePool.String()
+		case typesUtil.ActorType_Validator:
+			readyToUnstake, er = store.GetValidatorsReadyToUnstake(latestHeight, int32(typesUtil.StakeStatus_Unstaking))
+			poolName = typesUtil.PoolNames_ValidatorStakePool.String()
+		case typesUtil.ActorType_Undefined:
+			continue
 		}
 		if er != nil {
 			return typesUtil.ErrGetReadyToUnstake(er)
@@ -170,21 +162,22 @@ func (u *UtilityContext) UnstakeActorsThatAreReady() (err typesUtil.Error) {
 			if err = u.AddAccountAmountString(actor.GetOutputAddress(), actor.GetStakeAmount()); err != nil {
 				return err
 			}
-			if err = u.DeleteActor(utilActorType, actor.GetAddress()); err != nil {
-				return err
-			}
 		}
 	}
 	return nil
 }
 
 func (u *UtilityContext) BeginUnstakingMaxPaused() (err typesUtil.Error) {
-	latestHeight, err := u.GetLatestHeight()
+	latestHeight, err := u.GetLatestBlockHeight()
 	if err != nil {
 		return err
 	}
-	for _, UtilActorType := range UtilActorTypes {
-		maxPausedBlocks, err := u.GetMaxPausedBlocks(UtilActorType)
+	for _, actorTypeInt32 := range typesUtil.ActorType_value {
+		actorType := typesUtil.ActorType(actorTypeInt32)
+		if actorType == typesUtil.ActorType_Undefined {
+			continue
+		}
+		maxPausedBlocks, err := u.GetMaxPausedBlocks(actorType)
 		if err != nil {
 			return err
 		}
@@ -193,29 +186,29 @@ func (u *UtilityContext) BeginUnstakingMaxPaused() (err typesUtil.Error) {
 		if beforeHeight < 0 {
 			beforeHeight = 0
 		}
-		if err := u.UnstakeActorPausedBefore(beforeHeight, UtilActorType); err != nil {
+		if err := u.UnstakeActorPausedBefore(beforeHeight, actorType); err != nil {
 			return err
 		}
 	}
 	return nil
 }
 
-func (u *UtilityContext) UnstakeActorPausedBefore(pausedBeforeHeight int64, UtilActorType typesUtil.UtilActorType) (err typesUtil.Error) {
+func (u *UtilityContext) UnstakeActorPausedBefore(pausedBeforeHeight int64, ActorType typesUtil.ActorType) (err typesUtil.Error) {
 	var er error
 	store := u.Store()
-	unstakingHeight, err := u.GetUnstakingHeight(UtilActorType)
+	unstakingHeight, err := u.GetUnstakingHeight(ActorType)
 	if err != nil {
 		return err
 	}
-	switch UtilActorType {
-	case typesUtil.UtilActorType_App:
-		er = store.SetAppStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, typesUtil.UnstakingStatus)
-	case typesUtil.UtilActorType_Fish:
-		er = store.SetFishermanStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, typesUtil.UnstakingStatus)
-	case typesUtil.UtilActorType_Node:
-		er = store.SetServiceNodeStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, typesUtil.UnstakingStatus)
-	case typesUtil.UtilActorType_Val:
-		er = store.SetValidatorsStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, typesUtil.UnstakingStatus)
+	switch ActorType {
+	case typesUtil.ActorType_App:
+		er = store.SetAppStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, int32(typesUtil.StakeStatus_Unstaking))
+	case typesUtil.ActorType_Fisherman:
+		er = store.SetFishermanStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, int32(typesUtil.StakeStatus_Unstaking))
+	case typesUtil.ActorType_ServiceNode:
+		er = store.SetServiceNodeStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, int32(typesUtil.StakeStatus_Unstaking))
+	case typesUtil.ActorType_Validator:
+		er = store.SetValidatorsStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight, int32(typesUtil.StakeStatus_Unstaking))
 	}
 	if er != nil {
 		return typesUtil.ErrSetStatusPausedBefore(er, pausedBeforeHeight)
@@ -224,7 +217,7 @@ func (u *UtilityContext) UnstakeActorPausedBefore(pausedBeforeHeight int64, Util
 }
 
 func (u *UtilityContext) HandleProposalRewards(proposer []byte) typesUtil.Error {
-	feePoolName := typesGenesis.Pool_Names_FeeCollector.String()
+	feePoolName := typesUtil.PoolNames_FeeCollector.String()
 	feesAndRewardsCollected, err := u.GetPoolAmount(feePoolName)
 	if err != nil {
 		return err
@@ -248,7 +241,7 @@ func (u *UtilityContext) HandleProposalRewards(proposer []byte) typesUtil.Error 
 	if err = u.AddAccountAmount(proposer, amountToProposer); err != nil {
 		return err
 	}
-	if err = u.AddPoolAmount(typesGenesis.Pool_Names_DAO.String(), amountToDAO); err != nil {
+	if err = u.AddPoolAmount(typesUtil.PoolNames_DAO.String(), amountToDAO); err != nil {
 		return err
 	}
 	return nil
@@ -256,10 +249,9 @@ func (u *UtilityContext) HandleProposalRewards(proposer []byte) typesUtil.Error 
 
 // GetValidatorMissedBlocks gets the total blocks that a validator has not signed a certain window of time denominated by blocks
 func (u *UtilityContext) GetValidatorMissedBlocks(address []byte) (int, typesUtil.Error) {
-	store := u.Store()
-	height, er := store.GetHeight()
-	if er != nil {
-		return typesUtil.ZeroInt, typesUtil.ErrGetMissedBlocks(er)
+	store, height, err := u.GetStoreAndHeight()
+	if err != nil {
+		return 0, err
 	}
 	missedBlocks, er := store.GetValidatorMissedBlocks(address, height)
 	if er != nil {
@@ -282,29 +274,5 @@ func (u *UtilityContext) SetValidatorMissedBlocks(address []byte, missedBlocks i
 	if er != nil {
 		return typesUtil.ErrSetMissedBlocks(er)
 	}
-	return nil
-}
-
-func (u *UtilityContext) StoreBlock(blockProtoBytes []byte) error {
-	store := u.Store()
-
-	// Store in KV Store
-	if err := store.StoreBlock(blockProtoBytes); err != nil {
-		return err
-	}
-
-	// Store in SQL Store
-	// OPTIMIZE: Ideally we'd pass in the block proto struct to utility so we don't
-	//           have to unmarshal it here, but that's a major design decision for the interfaces.
-	codec := u.Codec()
-	block := &typesCons.Block{}
-	if err := codec.Unmarshal(blockProtoBytes, block); err != nil {
-		return typesUtil.ErrProtoUnmarshal(err)
-	}
-	header := block.BlockHeader
-	if err := store.InsertBlock(uint64(header.Height), header.Hash, header.ProposerAddress, header.QuorumCertificate); err != nil {
-		return err
-	}
-
 	return nil
 }