+++ utility/doc/CHANGELOG.md
@@ -7,6 +7,50 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.7] - 2022-10-12
+
+### [#235](https://github.com/pokt-network/pocket/pull/235) Config and genesis handling
+
+- Updated to use `RuntimeMgr`
+- Made `UtilityModule` struct unexported
+- Updated tests and mocks
+- Removed some cross-module dependencies
+
+## [0.0.0.6] - 2022-10-06
+
+- Don't ignore the exit code of `m.Run()` in the unit tests
+- Fixed several broken unit tests related to type casting
+- Removed some unit tests (e.g. `TestUtilityContext_UnstakesPausedBefore`) that were legacy and replaced by more general ones (e.g. `TestUtilityContext_UnstakePausedBefore`)
+- Avoid exporting unnecessary test helpers
+
+## [0.0.0.5] - 2022-09-29
+
+- Remove unused `StoreBlock` function from the utility module interface
+
+## [0.0.0.4] - 2022-09-23
+
+- Created `UtilityConfig`
+- Added `max_mempool_transaction_bytes` and `max_mempool_transactions` to the utility
+  config to allow dynamic configuration of the mempool
+- Matched configuration unmarshalling pattern of other modules
+- Added V0 mempool default configurations
+- Regenerated build files with new mempool config
+- Consolidated `UtilActorType` in `utility` and `utility/types` to `typesUtil.ActorType`
+- Deprecated all code in `actor.go` and replaced with test helpers
+- Converted stake status to proto.enum (int32)
+- Added DISCUSS items around shared code and `StakeStatus`
+- Removed no-op `DeleteActor` code
+- Improved unit test for `UnstakeActorsThatAreReady()`
+- Removed all usages of `fmt.Sprintf()` from the testing package
+- Replaced all usages of `requre.True/require.False` with `require.Equal` unless checking a boolean
+- Added helper function for getting height and store for a readable and consistent `typesUtil.Error` value
+- Added testing.M argument to `newTestingPersistenceModule`
+- Moved in-function _literal_ arguments for `newTestingPersistenceModule` to private constants
+- Added the address parameter to `ErrInsufficientFunds` function for easier debugging
+- Added unit test for `LegacyVote.ValidateBasic()`
+- Added `ErrUnknownActorType` to all switch statements on `actorType`
+- Removed `import` of `consTypes` (consensus module)
+
 ## [0.0.0.3] - 2022-09-15
 
 ### Code cleanup